richat-climatique/backend/requirements.txt
Django==4.2.7 
djangorestframework==3.14.0 
django-cors-headers==4.3.1 
python-decouple==3.8 
Pillow==10.0.1 
django-filter==23.3 
mysqlclient==2.2.0 
